+++ api/container_delete.go
@@ -15,10 +15,10 @@ func (c *API) ContainerDelete(ctx context.Context, name string, force bool, dele
 		return err
 	}
 
-	defer res.Body.Close()
+	defer ignoreClose(res.Body)
 
 	if res.StatusCode == http.StatusNoContent {
 		return nil
 	}
-	return fmt.Errorf("unknown error, status code: %d", res.StatusCode)
+	return fmt.Errorf("cannot delete container, status code: %d", res.StatusCode)
 }