+++ api/container_start.go
@@ -3,7 +3,7 @@ package api
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"time"
 )
@@ -16,11 +16,11 @@ func (c *API) ContainerStart(ctx context.Context, name string) error {
 		return err
 	}
 
-	defer res.Body.Close()
+	defer ignoreClose(res.Body)
 
 	if res.StatusCode != http.StatusNoContent {
-		body, _ := ioutil.ReadAll(res.Body)
-		return fmt.Errorf("unknown error, status code: %d: %s", res.StatusCode, body)
+		body, _ := io.ReadAll(res.Body)
+		return fmt.Errorf("cannot start container, status code: %d: %s", res.StatusCode, body)
 	}
 
 	// wait max 10 seconds for running state