+++ api/image_pull.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 )
 
@@ -29,10 +28,10 @@ func (c *API) ImagePull(ctx context.Context, nameWithTag string, auth ImageAuthC
 		return "", err
 	}
 
-	defer res.Body.Close()
+	defer ignoreClose(res.Body)
 	if res.StatusCode != http.StatusOK {
-		body, _ := ioutil.ReadAll(res.Body)
-		return "", fmt.Errorf("unknown error, status code: %d: %s", res.StatusCode, body)
+		body, _ := io.ReadAll(res.Body)
+		return "", fmt.Errorf("cannot pull image, status code: %d: %s", res.StatusCode, body)
 	}
 
 	dec := json.NewDecoder(res.Body)