+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	log "github.com/hashicorp/go-hclog"
+	"github.com/hashicorp/go-hclog"
 	_ "github.com/hashicorp/nomad/api"
 	"github.com/hashicorp/nomad/plugins"
 )
@@ -12,6 +12,6 @@ func main() {
 }
 
 // factory returns a new instance of the LXC driver plugin
-func factory(log log.Logger) interface{} {
+func factory(log hclog.Logger) interface{} {
 	return NewPodmanDriver(log)
 }